Sync upstream (#96)

+++ src/main/java/com/cleanroommc/modularui/utils/DAM.java
@@ -266,7 +266,7 @@ public class DAM {
     }
 
     public static double[] pow(double[] src, double op, double @Nullable [] res) {
-        return applyEach(src, v -> (double) Math.pow(v, op), res);
+        return applyEach(src, v -> Math.pow(v, op), res);
     }
 
     public static double[] diff(double[] src) {
@@ -342,7 +342,7 @@ public class DAM {
             y += coeff[0];
             if (coeff.length == 2) return y + x * coeff[1];
             for (int i = 1; i < coeff.length; i++) {
-                y += (double) (Math.pow(x, i) * coeff[i]);
+                y += Math.pow(x, i) * coeff[i];
             }
             return y;
         }, res);
@@ -380,7 +380,7 @@ public class DAM {
     }
 
     public static double geometricMean(double[] src) {
-        return (double) Math.pow(product(src), 1f / src.length);
+        return Math.pow(product(src), 1f / src.length);
     }
 
     public static double[] concat(double[] a, double[] b) {
